Fix GitHub Pages build & deploy
Previously we had `actions/configure-pages` in the wrong CI job,
though it's slightly unclear what impact that had. We also were
still using an older version of `actions/deploy-pages`, which
seems to have been missed in https://github.com/srobo/srawn/pull/80
and which doesn't seem to be compatible with the artifact actions
version we're now using.